mid: 0000528 msg: Merge de la branca 2.6
git-svn-id: svn://svn.code.sf.net/p/gehelium/code/helium/branches/helium-3.0@858 e3271b2b-fd41-4d56-bedf-4b5636213925

+++ helium-core/src/main/java/net/conselldemallorca/helium/core/model/dao/HibernateGenericDao.java
@@ -33,8 +33,6 @@ public class HibernateGenericDao<T, ID extends Serializable> extends HibernateDa
 	
 	private NamedParameterJdbcTemplate jdbcTemplate;
 
-
-
 	public HibernateGenericDao() {
 	}
 
@@ -220,7 +218,7 @@ public class HibernateGenericDao<T, ID extends Serializable> extends HibernateDa
 	
 	public int getCountByCriteria(final Criteria crit) {
 		crit.setProjection(Projections.rowCount());
-		final Integer result = (Integer)crit.uniqueResult();
+		final Long result = (Long)crit.uniqueResult();
 		return (result == null) ? 0 : result.intValue();
 	}
